requirements.txt
bluepy==1.3.0
cantools==37.2.0
python_can==3.0.0
persist_queue==0.8.0
protobuf==4.21.8
pynmea2==1.18.0
pyserial==3.5
websockets==10.3
mypy==0.982
mypy-protobuf==3.4.0
paho-mqtt==1.6.1
